+++ tests/CMakeLists.txt
@@ -42,8 +42,9 @@ set(UNITTEST_SOURCES
     "csv_routes_csr_tests.cpp" 
     "vertexlist_tests.cpp" 
     "incidence_tests.cpp"  
-    "neighbors_tests.cpp"  
+    "neighbors_tests.cpp"
     "edgelist_tests.cpp" 
+    "edgelist_view_tests.cpp" 
     "examples_tests.cpp"
     "shortest_paths_tests.cpp" 
     "transitive_closure_tests.cpp"
